Fix slurm maintanance jobs (#236)
* Eliminate slurm calls for maintanance operations

Slurm jobs cannot be utilzed to perform maintanance tasks such as checking image
availability and node health as this causes livelock and inefficient resource usage.
This PR eliminates all srun/sbatch calls except for the one submitting scarab jobs.
Checking node availability can be performed locally by checking "sinfo -Nel". Docker
image availablity check (and potential pull) is performed as part of every job
submission (we now submit a script that atomically checks image availability and
then submits the job). Submitting jobs is now non-blocking, as such, we cannot run
finish_sim to i.e. perform stats collection. This needs to be performed as part of
the visualize step. Also fixes a bug in status to determine running jobs.
This PR improves resource utiliation and overall simulation speed by 5x.

Limitations:
- The new script assumes that any node that is not in "drain" state has docker daemon
running. The proper way to ensure this is have e.g. a cron job that ensures this is
the case when a node is added to idle state (TODO).
- Not tested with trace mode

* --status broke due to new lines added to logfile when prepare_docker_image needed to build container. Added preamble/epilouge (BEGIN/END) to prepare_docker_image.py output, and ignored these lines in --status error checking

* Added more specific docker image preperation error handling

* 4 Fixes: (1) Count submitted jobs correctly (2) Create log file on submit so taks are pending not non-existing (3) fix line used to determine running (4) extract right line for configs

* Fix accounting for running, pending, completed, total

---------

Co-authored-by: Alexander Symons <aesymons888@gmail.com>